+++ examples/collect-card-details/CHANGELOG.md
@@ -1,5 +1,11 @@
 # example-ui-components
 
+## 0.2.4
+
+### Patch Changes
+
+- @evervault/browser@2.37.2
+
 ## 0.2.3
 
 ### Patch Changes